split ImsError into specific classes for better testing
