+++ src/models/errors/ims-server-connection-error.ts
@@ -1,0 +1,8 @@
+import { ImsError } from './ims-error';
+export class ImsServerConnectionError extends ImsError {
+
+  constructor(server: string, err: any) {
+    super('Verbindung zum IMS Rest Server ' + server + ' nicht möglich.', err);
+    Object.setPrototypeOf(this, ImsServerConnectionError.prototype);
+  }
+}
