+++ src/pages/login/login.spec.ts
@@ -1,4 +1,5 @@
-import { ImsError } from './../../models/ims-error';
+import { ImsAuthenticationError } from './../../models/errors/ims-authentication-error';
+import { ImsServerConnectionError } from './../../models/errors/ims-server-connection-error';
 import { Info } from './../../models/info';
 import { SettingArchivePage } from './../setting-archive/setting-archive';
 import { TestBed, inject, async, ComponentFixture } from '@angular/core/testing';
@@ -65,26 +66,24 @@ describe('Page: Login', () => {
     expect(toastController.create).toHaveBeenCalled();
   }));
 
-  it('Show and Hide loading with alert on wrong server', inject([LoadingService, AlertService], (loadingService: LoadingService, alertService: AlertService) => {
+  it('Show and Hide loading with error thrown on wrong server', inject([LoadingService], (loadingService: LoadingService) => {
     spyOn(loadingService, 'showLoading').and.callThrough();
     spyOn(loadingService, 'hideLoading').and.callThrough();
-    spyOn(alertService, 'showError').and.callThrough();
     page.loginForm.controls['server'].setValue('wrong');
     page.loginForm.controls['user'].setValue('wrong');
     page.loginForm.controls['password'].setValue('wrong');
-    page.login();
+    expect(() => page.login()).toThrowError(ImsServerConnectionError);
     expect(loadingService.showLoading).toHaveBeenCalledTimes(1);
     expect(loadingService.hideLoading).toHaveBeenCalledTimes(1);
-    expect(alertService.showError).toHaveBeenCalled();
   }));
 
- it('Show and Hide loading with error thrown on wrong username', inject([LoadingService, AlertService, ImsBackendMock], (loadingService: LoadingService, alertService: AlertService, imsBackendMock: ImsBackendMock) => {
+ it('Show and Hide loading with error thrown on wrong username', inject([LoadingService, ImsBackendMock], (loadingService: LoadingService, imsBackendMock: ImsBackendMock) => {
     spyOn(loadingService, 'showLoading').and.callThrough();
     spyOn(loadingService, 'hideLoading').and.callThrough();
     page.loginForm.controls['server'].setValue(imsBackendMock.credential.server);
     page.loginForm.controls['user'].setValue('wrong');
     page.loginForm.controls['password'].setValue('wrong');
-    expect(() => page.login()).toThrowError(ImsError);
+    expect(() => page.login()).toThrowError(ImsAuthenticationError);
     expect(loadingService.showLoading).toHaveBeenCalledTimes(1);
     expect(loadingService.hideLoading).toHaveBeenCalledTimes(1);
   }));
